+++ docs/changelog.rst
@@ -9,6 +9,8 @@ v1.1.2
   * Fixed target name deduction with chained function calls on
     generator functions (#122)
   * MSVC fixed handling of keep_suffix and force_suffix options
+  * MSVC compiler now correctly deduces a correct compiler version
+    if multiple versions are specified and the first can not be used
 
 * Behaviour changes